backend/requirements.txt
Flask
zai-sdk
flask-cors
python-dotenv